Move notify to a library directory.
